requirements/production.txt
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.2.0
cffi==1.14.3
colorama==0.4.3
cryptography==3.1.1
Django==3.1
django-crispy-forms==1.9.2
Fabric3==1.14.post1
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
paramiko==2.7.2
pycparser==2.20
pylint==2.6.0
PyNaCl==1.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1